+++ sms/__init__.py
@@ -5,7 +5,7 @@ import os
 
 
 # Define Shared Constants
-__version__ = "0.0.9-5"
+__version__ = "0.0.9-7"
 """str: Application version number."""