+++ sms/types/action.py
@@ -6,6 +6,10 @@ from enum import auto, Enum
 
 __all__ = (
         'ActType',
+        'NORMAL_ACTS',
+        'OBJECT_ACTS',
+        'NO_SUBJECT_ACTS',
+        'FALG_ACTS',
         )
 
 
@@ -28,6 +32,8 @@ class ActType(Enum):
     # view
     DRAW = auto()
     EXPLAIN = auto()
+    SKY = auto()
+    LIGHT = auto()
     # dialogue
     TALK = auto()
     THINK = auto()
@@ -56,6 +62,7 @@ class ActType(Enum):
                 ActType.FORESHADOW: ('foreshadow', 'FS'),
                 ActType.GO: ('go'),
                 ActType.HAVE: ('have'),
+                ActType.LIGHT: ('light',),
                 ActType.MARK: ('mark'),
                 ActType.NONE: ('none'),
                 ActType.NOTE: ('note'),
@@ -64,6 +71,7 @@ class ActType(Enum):
                 ActType.PUT: ('put'),
                 ActType.RID: ('rid'),
                 ActType.SAME: ('same', '-'),
+                ActType.SKY: ('sky'),
                 ActType.TALK: ('talk'),
                 ActType.THINK: ('think'),
                 ActType.TITLE: ('title'),
@@ -72,7 +80,7 @@ class ActType(Enum):
                 }[self]
 
 
-NORMAL_ACTION = [
+NORMAL_ACTS = [
         ActType.BE,
         ActType.COME,
         ActType.DO,
@@ -91,4 +99,18 @@ NORMAL_ACTION = [
 OBJECT_ACTS = [
         ActType.PUT,
         ActType.RID,
+        ]
+
+
+NO_SUBJECT_ACTS = [
+        ActType.SKY,
+        ActType.LIGHT,
+        ActType.NOTE,
+        ActType.PLOT,
+        ]
+
+
+FLAG_ACTS = [
+        ActType.FORESHADOW,
+        ActType.PAYOFF,
         ]